Fixed functions in all three projects with the same fixes
Altered commitAndGetSale() so it now returns the Id as well.

Altered putUpForSale() to properly set the currentPurchasePrice once a new memberCartItem
is created.

EXTRA:

Modified checkoutController so it no longer tries to empty the shoppingCart at the end a purchase
as it should already be empty. That should fix the problem that console showed an error message that
the shopping cart failed to be emptied once a clientOrder was concluded.